1. change 'skip to config' button to L4SLBWizard.tsx file since easy to support edit 2. refactory SLBConfig code    2.1 change most of api call into parent level component L4SLBConfig    2.2 move all generate name functions into final format functions 3. comment out all utility class functions
