+++ hc-gui/src/containers/L4SLB/L4SLBWizard/Steps/DeploymentForm/DeploymentForm.tsx
@@ -228,9 +228,9 @@ export default class DeploymentForm extends AbstractStep<
           <A10Button className="btn-next" type="primary" onClick={this.onNext}>
             Next
           </A10Button>
-          <A10Button className="btn-action">
+          {/* <A10Button className="btn-action">
             <Link to="/configuration">Skip Wizard to configuration</Link>
-          </A10Button>
+          </A10Button> */}
         </div>
       </React.Fragment>
     )
