+++ hc-gui/src/containers/L4SLB/SLBConfiguration/VirtualServer/index.tsx
@@ -1,0 +1,1 @@
+export {default,default as SLBConfigurationForm} from './SLBConfiguration'
